Fix call parameter parsing, fix handling of empty strings in the lexer

+++ src/main/kotlin/dev/vanadium/avo/runtime/interpreter/types/DataType.kt
@@ -21,7 +21,7 @@ sealed class DataType {
 
     object FloatType : DataType(), KTypeMappable{
         override fun toString() = "Float"
-        override fun toKType(): KType = Float::class.createType()
+        override fun toKType(): KType = Double::class.createType()
     }
 
     object StringType : DataType(), KTypeMappable {
